requirements.txt
lxml==4.5.0
Pillow==7.0.0
python-pptx==0.6.18
XlsxWriter==1.2.8
